backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
fonttools==4.54.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
icecream==2.1.3
idna==3.10
imageio==2.35.1
install==1.3.5
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
math123==0.4
matplotlib==3.9.2
matplotlib-inline==0.1.7
mysqlclient==2.2.4
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.1
outcome==1.3.0.post0
packaging==24.0
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.10
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==1.7.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyzmq==26.0.3
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.31
sqlparse==0.5.1
stack-data==0.6.3
tifffile==2024.9.20
tornado==6.4
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.26.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
